Floating Points (#2)
Replaced all default 'Float' types with 'Double'.

+++ Sources/GraphPoint/Radius.swift
@@ -1,4 +1,4 @@
 import Foundation
 
 /// The radius of a circle is the length of the line from the center to any point on its edge.
-public typealias Radius = Float
+public typealias Radius = Double
